# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "jump.rs",
    "description": "Wake-on-LAN API for managing and waking network devices",
    "license": {
      "name": "MIT"
    },
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Get current authenticated user",
        "description": "Returns information about the currently authenticated user.",
        "operationId": "me",
        "responses": {
          "200": {
            "description": "Current user info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "basic_auth": []
          }
        ]
      }
    },
...
    "/api/devices": {
      "get": {
        "tags": [
          "devices"
        ],
        "summary": "List all devices",
        "description": "Returns a list of all registered devices that can receive Wake-on-LAN packets.",
        "operationId": "getDevices",
        "responses": {
          "200": {
            "description": "List of all devices",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Device"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "devices"
        ],
        "summary": "Create a new device",
        "description": "Creates a new device that can receive Wake-on-LAN packets. The MAC address must be valid.",
        "operationId": "createDevice",
        "requestBody": {
          "description": "Device to create",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDeviceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Device created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Device"
                }
              }
            }
          },
          "400": {
            "description": "Validation error (e.g., invalid MAC address)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateDeviceRequest": {
        "type": "object",
        "required": [
          "name",
          "mac_address"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ],
            "example": "My main gaming rig"
          },
          "ip_address": {
            "type": [
              "string",
              "null"
            ],
            "example": "192.168.1.100"
          },
          "mac_address": {
            "type": "string",
            "example": "00:11:22:33:44:55"
          },
          "name": {
            "type": "string",
            "example": "Gaming PC"
          },
          "port": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "example": 9,
            "minimum": 0
          }
        }
      },
      "Device": {
        "type": "object",
        "required": [
          "id",
          "name",
          "mac_address",
          "port",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "example": "My main computer"
          },
          "id": {
            "type": "string",
            "example": "V1StGXR8_Z5jdHi6B"
          },
          "ip_address": {
            "type": [
              "string",
              "null"
            ],
            "example": "192.168.1.100"
          },
          "mac_address": {
            "type": "string",
            "example": "00:11:22:33:44:55"
          },
          "name": {
            "type": "string",
            "example": "Gaming PC"
          },
          "port": {
            "type": "integer",
            "format": "int32",
            "default": 9,
            "minimum": 0
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error response returned by all API endpoints",
        "required": [
          "status",
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "description": "Human-readable error message",
            "example": "Invalid MAC address format: xyz"
          },
          "status": {
            "type": "string",
            "description": "Always \"error\" for error responses",
            "example": "error"
          }
        }
      },
...
      "MeResponse": {
        "type": "object",
        "required": [
          "username"
        ],
        "properties": {
          "username": {
            "type": "string",
            "example": "admin"
          }
        }
      },
...
    }
  },
...
}